@@ -7,7 +7,7 @@
     "build": "gatsby build",
     "develop": "gatsby develop",
     "serve": "gatsby serve",
-    "lint": "eslint --ext .js,.jsx,.ts,.tsx src && tsc --noEmit",
+    "lint": "eslint src && tsc --noEmit",
     "format": "prettier --write 'src/**/*.ts(x)'"
   },
   "dependencies": {
@@ -32,6 +32,7 @@
     "slick-carousel": "^1.8.1"
   },
   "devDependencies": {
+    "@eslint/js": "^9.3.0",
     "@types/node": "20.12.11",
     "@types/react": "18.3.2",
     "@types/react-dom": "18.3.0",
@@ -39,23 +40,14 @@
     "@types/react-slick": "^0",
     "@types/react-transition-group": "4.4.10",
     "@types/slick-carousel": "^1",
-    "@typescript-eslint/eslint-plugin": "7.8.0",
-    "@typescript-eslint/parser": "7.8.0",
-    "eslint": "9.2.0",
-    "eslint-config-airbnb": "19.0.4",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "3.6.1",
-    "eslint-plugin-import": "2.29.1",
+    "eslint": "^9.3.0",
     "eslint-plugin-jsx-a11y": "6.8.0",
-    "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-promise": "6.1.1",
-    "eslint-plugin-react": "7.34.1",
-    "eslint-plugin-react-hooks": "4.6.2",
     "gatsby-plugin-graphql-codegen": "3.1.1",
     "graphql": "16.8.1",
     "prettier": "3.2.5",
-    "typescript": "5.4.5"
+    "typescript": "^5.4.5",
+    "typescript-eslint": "^7.9.0"
   },
   "repository": {
     "type": "git",
